# Introduce server-side TTL cache in `GrpcStorageProxy`