--- a/optuna/storages/_grpc/servicer.py
+++ b/optuna/storages/_grpc/servicer.py
@@ -1,8 +1,11 @@
 from __future__ import annotations
 
+from collections import defaultdict
+import dataclasses
 from datetime import datetime
 import json
 import threading
+import time
 from typing import TYPE_CHECKING
 
 from optuna import logging
@@ -32,10 +35,18 @@ _logger = logging.get_logger(__name__)
 DATETIME_FORMAT = "%Y-%m-%d %H:%M:%S.%f"
 
 
+@dataclasses.dataclass
+class CachedTrials:
+    trials: list[FrozenTrial] = dataclasses.field(default_factory=list)
+    last_sync_time: float = 0.0
+
+
 class OptunaStorageProxyService(api_pb2_grpc.StorageServiceServicer):
-    def __init__(self, storage: BaseStorage) -> None:
+    def __init__(self, storage: BaseStorage, ttl_seconds: float = 0.0) -> None:
         self._backend = storage
         self._lock = threading.Lock()
+        self._ttl_seconds = ttl_seconds
+        self._cached_trials: defaultdict[int, CachedTrials] = defaultdict(CachedTrials)
 
     def CreateNewStudy(
         self,
@@ -329,14 +340,17 @@ class OptunaStorageProxyService(api_pb2_grpc.StorageServiceServicer):
         study_id = request.study_id
         included_trial_ids = set(request.included_trial_ids)
         trial_id_greater_than = request.trial_id_greater_than
-        try:
-            trials = self._backend.get_all_trials(study_id, deepcopy=False)
-        except KeyError as e:
-            context.abort(code=grpc.StatusCode.NOT_FOUND, details=str(e))
-
+        if time.time() - self._cached_trials[study_id].last_sync_time >= self._ttl_seconds:
+            try:
+                self._cached_trials[study_id].trials = self._backend.get_all_trials(
+                    study_id, deepcopy=False
+                )
+            except KeyError as e:
+                context.abort(code=grpc.StatusCode.NOT_FOUND, details=str(e))
+            self._cached_trials[study_id].last_sync_time = time.time()
         filtered_trials = [
             _to_proto_trial(t)
-            for t in trials
+            for t in self._cached_trials[study_id].trials
             if t._trial_id > trial_id_greater_than or t._trial_id in included_trial_ids
         ]
         return api_pb2.GetTrialsReply(trials=filtered_trials)