<!-- Thank you for creating a pull request! In general, we merge your pull request after it gets two or more approvals. To proceed to the review process by the maintainers, please make sure that the PR meets the following conditions: (1) it passes all CI checks, and (2) it is neither in the draft nor WIP state. If you wish to discuss the PR in the draft state or need any other help, please mention the Optuna development team in the PR. -->

## Motivation
To make `GrpcStorageProxy` faster by introducing server-side TTL cache.
In the gRPC storage proxy, if a client reads get_all_trials continuously for a short period of time (e.g. 1 sec), the state of the RDB is expected to not change.
TTL cache for trials can be introduced on the server side to speed up the process. 

## Description of the changes
This PR makes `GrpcStorageProxy` faster by cache `OptunaStorageProxyService.GetTrials` calls. Here is the performance comparison results and the benchmark script I used.

### Results
|  | `master` branch | This PR (ttl_cache_seconds = 10) |
| --- | ------------- | ------------- |
| total | 859.9sec | 233.1sec |
| `client.py:368(get_all_trials)` | 705.3sec | 16.83sec |
| stats | ![image](https://github.com/user-attachments/assets/b4b541b1-29dd-4842-84b3-292a10570319) | ![image](https://github.com/user-attachments/assets/fb8b4f29-ed99-4541-b9d4-5da4cb0ac355) |

<details>
<summary>client.py</summary>

```python
from __future__ import annotations

import argparse
import cProfile
import time
import optuna
from typing import Literal
from optuna.storages import GrpcStorageProxy, BaseStorage
from time import sleep
from uuid import uuid4


study_name = "ttl" + str(uuid4())[:8]
n_trials = 1000

STORAGE_KIND = Literal["grpc", "rdb", "middleware"]
STORAGE_URL = "YOUR_STORAGE_URL"


def get_storage(storage_kind: STORAGE_KIND) -> BaseStorage:
    if storage_kind == "grpc":
        return GrpcStorageProxy(host="localhost", port=13000)
    elif storage_kind == "rdb":
        return optuna.storages.get_storage(STORAGE_URL)


def objective(trial: optuna.Trial) -> float:
    s = 0.0
    for i in range(10):
        trial.set_user_attr(f"attr{i}", "attr value")
    s += trial.suggest_float("x", -10, 10) ** 2
    return s


def study_optimize(storage_kind: STORAGE_KIND, n_trials: int) -> int:
    start = time.time()
    storage = get_storage(storage_kind)
    while True:
        try:
            study = optuna.load_study(storage=storage, study_name=study_name, sampler=optuna.samplers.RandomSampler())
            break
        except:
            sleep(10)
    study.optimize(objective, callbacks=[optuna.study.MaxTrialsCallback(n_trials)])
    elapsed = time.time() - start
    return elapsed


def bench_study_optimize(storage_kind: STORAGE_KIND) -> list[int]:
    elapsed = []
    _study = optuna.create_study(
        storage=get_storage("rdb"),
        study_name=study_name,
        load_if_exists=True,
        sampler=optuna.samplers.RandomSampler(),
    )
    study_optimize(storage_kind, n_trials)
    return elapsed


def main(storage: STORAGE_KIND, benchmark_file: str):
    try:
        optuna.delete_study(storage=get_storage("rdb"), study_name=study_name)
    except:
        pass
    cProfile.runctx(
        f"bench_study_optimize('{storage}')",
        globals(),
        locals(),
        filename=benchmark_file,
    )


if __name__ == "__main__":
    parser = argparse.ArgumentParser()
    parser.add_argument("--storage", type=str, required=True)
    parser.add_argument("--benchmark_file", type=str, required=True)
    args = parser.parse_args()
    main(args.storage, args.benchmark_file)
```

</details>

<details>
<summary>server.py</summary>

```
from __future__ import annotations

import argparse
from optuna.storages import run_grpc_proxy_server
import optuna
import signal


STORAGE_URL = "YOUR_STORAGE_URL"

def handler(signum, frame):
    print("Received signal", signum)
    exit(0)

if __name__ == "__main__":
    parser = argparse.ArgumentParser()
    parser.add_argument("--ttl-cache-seconds", type=float, default=None)
    args = parser.parse_args()
    signal.signal(signal.SIGINT, handler)
    signal.signal(signal.SIGTERM, handler)

    run_grpc_proxy_server(
        optuna.storages.RDBStorage(STORAGE_URL),
        host="localhost",
        port=13000,
        ttl_cache_seconds=args.ttl_cache_seconds,
    )
```

</details>