@HideakiImamura 2025-03-28
Thanks for the PR! The change basically looks good to me. Could you tell me the value of `ttl_cache_seconds` in the benchmark comparison?