@hitsgub 2025-04-17
The file for test added in this PR will also be changed according to the following PR.

https://github.com/optuna/optuna/pull/6044#discussion_r2046187267