@hitsgub 2025-04-25
完了したstudyがちゃんと消えることを確認
なんらかの理由で途中でworkflowを潰した場合は…放置するか手動でなんとかするしかないですね