@hitsgub 2025-05-02
When ttl cache is enabled and parallel processing is used, it has been observed that multiple processes may generate trials with the same trial_id.

The fundamental solution is:
• Stop using sequential trial_ids and instead use UUIDs, then rewrite all trial_id ordering-dependent logic to use timestamps.

As a provisional solution:
• When calling get_all_trials within create_new_trial, force synchronization with the server by ignoring the ttl cache.